Remove hotspots
